Updates makefile to use opam for dependencies.
After this commit, one can install all the dependencies by running

  make depends

This temporarily pins b2piet as an opam package, then installs all the
dependencies, without installing b2piet.

The main advantage of doing it this way is that opam will check the
compiler version, and it is possible to add version constraints on the
packages later.

The commit also lists package conf-libpng as a dependency. Installing
conf-libpng runs a check for the presence of libpng on the user's
system. If libpng is not installed, the user gets an error message with
instructions.

I am concerned that there is not an optional dependency from camlimages
on conf-libpng, which could, in principle, permit camlimages to be
compiled before the conf-libpng check is run. If that really is
possible, the problem should be fixed upstream in the opam repository.

I didn't update the README.